change date_st members to int from long, so that it remains the same size as it has always been even on systems with 64-bit longs, and since the fields do not need to be any bigger than that
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/rdsslibs@8272 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ ui/ui_param.h
@@ -27,8 +27,8 @@ typedef unsigned char byte;	/* Basic byte variable */
 
 struct date_st
 {
-	long	ds_yymmdd;	/* Day portion	*/
-	long	ds_hhmmss;	/* time portion */
+  int	ds_yymmdd;	/* Day portion	*/
+  int	ds_hhmmss;	/* time portion */
 };
 typedef struct date_st date;		/* Date in date/time format */
